Move game logic to game class and added new loop to controller
