+++ CodeSmells/GameController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.AccessControl;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,124 +9,42 @@ namespace CodeSmells
 {
     internal class GameController
     {
-        private IUI ui;
+        private IUI uI;
+        private IGame game;
         public GameController(IUI ui) 
         {
-            this.ui = ui;
+            this.uI = ui;
+        }
+        private void SetGame(IGame game)
+        {
+            this.game = game;
         }
 
         public void StartGame()
         {
-            bool keepPlaying = true;
-            ui.PutString("Enter your user name:\n");
-            string playerName = ui.GetString();
-
-            while (keepPlaying)
+            do
             {
-                string gameGoal = GenerateRandomNumber();
-
-
-                ui.PutString("New game:\n");
-                //comment out or remove next line to play real games!
-                ui.PutString("For practice, number is: " + gameGoal + "\n");
-                string playerGuess = ui.GetString();
-
-                int numberOfPlayerGuesses = 1;
-                string bullsAndCowsResult = CompareGuessToGoal(gameGoal, playerGuess);
-                ui.PutString(bullsAndCowsResult + "\n");
-                while (bullsAndCowsResult != "BBBB,")
+                uI.PutString(
+                "Choose one:\n" +
+                "1: Moo\n" +
+                "2: Mastermind\n" +
+                "3: Exit");
+                int gameChoice = Convert.ToInt32(uI.GetString());
+                switch (gameChoice)
                 {
-                    numberOfPlayerGuesses++;
-                    playerGuess = ui.GetString();
-                    ui.PutString(playerGuess + "\n");
-                    bullsAndCowsResult = CompareGuessToGoal(gameGoal, playerGuess);
-                    ui.PutString(bullsAndCowsResult + "\n");
-                }
-                StreamWriter output = new StreamWriter("result.txt", append: true);
-                output.WriteLine(playerName + "#&#" + numberOfPlayerGuesses);
-                output.Close();
-                DisplayPlayerTopList();
-                ui.PutString("Correct, it took " + numberOfPlayerGuesses + " guesses\nContinue?");
-                string answer = ui.GetString();
-                if (answer != null && answer != "" && answer.Substring(0, 1) == "n")
-                {
-                    keepPlaying = false;
-                }
-            }
-        }
-        static string GenerateRandomNumber()
-        {
-            Random randomGenerator = new Random();
-            string randomNumbers = "";
-            for (int i = 0; i < 4; i++)
-            {
-                int random = randomGenerator.Next(10);
-                string randomDigit = "" + random;
-                while (randomNumbers.Contains(randomDigit))
-                {
-                    random = randomGenerator.Next(10);
-                    randomDigit = "" + random;
-                }
-                randomNumbers = randomNumbers + randomDigit;
-            }
-            return randomNumbers;
-        }
+                    case 1:
+                        SetGame(new Moo(uI));
+                        break;
+                    case 2:
+                        //SetGame(new Mastermind(uI));
+                        break;
+                    case 3:
+                        uI.Exit();
+                        break;
 
-        static string CompareGuessToGoal(string goal, string guess)
-        {
-            int cows = 0, bulls = 0;
-            guess += "    ";     // if player entered less than 4 chars
-            for (int i = 0; i < 4; i++)
-            {
-                for (int j = 0; j < 4; j++)
-                {
-                    if (goal[i] == guess[j])
-                    {
-                        if (i == j)
-                        {
-                            bulls++;
-                        }
-                        else
-                        {
-                            cows++;
-                        }
-                    }
                 }
-            }
-            return "BBBB".Substring(0, bulls) + "," + "CCCC".Substring(0, cows);
-        }
-
-
-        public void DisplayPlayerTopList()
-        {
-            StreamReader input = new StreamReader("result.txt");
-            List<Player> playerResults = new List<Player>();
-            string line;
-            while ((line = input.ReadLine()) != null)
-            {
-                string[] nameAndScore = line.Split(new string[] { "#&#" }, StringSplitOptions.None);
-                string name = nameAndScore[0];
-                int guesses = Convert.ToInt32(nameAndScore[1]);
-                Player player = new Player(name, guesses);
-                int pos = playerResults.IndexOf(player);
-                if (pos < 0)
-                {
-                    playerResults.Add(player);
-                }
-                else
-                {
-                    playerResults[pos].Update(guesses);
-                }
-
-
-            }
-            playerResults.Sort((p1, p2) => p1.Average().CompareTo(p2.Average()));
-            ui.PutString("Player   games average");
-            foreach (Player p in playerResults)
-            {
-                ui.PutString(string.Format("{0,-9}{1,5:D}{2,9:F2}", p.Name, p.NGames, p.Average()));
-            }
-            input.Close();
-        }
+                game.PlayGame();
+            } while (true);
+        }        
     }
 }
